Add `sibasevalue` function
* Define uexpand for quantities with Dimensions
* Define `sibasevalue(q)=ustrip(uexpand(q))

When obtaining the numerical value of a quantity (e.g. for computations
with sparse direct solvers implemented in other languages) via `ustrip`,
there is danger of mixing up the right values when quantities with
symbolic dimensions come into play.

By adding a `uexpand` method for quantities with dimensions,
which just passes through the value one can write code consistent
for both flavors of quantities.

`sibasevalue` shall help to remove a little bit of boilerplate
and explicitely hint at the characteristics of the returned value.